r3k-bdttools-master/requirements.txt
numpy
awkward
uproot
scikit-learn
xgboost
tqdm
matplotlib
pyyaml